--- routes/__layout.svelte
@@ -1,16 +1,0 @@
-<script lang="ts">
-  import {navigating} from '$app/stores';
-  import NProgress from 'nprogress';
-
-  $: if ($navigating) {
-    NProgress.start();
-  } else {
-    NProgress.done();
-  }
-
-  import "../app.css";
-</script>
-
-<template>
-  <slot />
-</template>
