+++ src/routes/__layout.svelte
@@ -1,5 +1,16 @@
-<script>
+<script lang="ts">
+  import {navigating} from '$app/stores';
+  import NProgress from 'nprogress';
+
+  $: if ($navigating) {
+    NProgress.start();
+  } else {
+    NProgress.done();
+  }
+
   import "../app.css";
 </script>
 
-<slot />
+<template>
+  <slot />
+</template>
